move api/pki.go ro pki_list.go
because it's somewhat complicated.
