+++ api/pki.go
@@ -15,18 +15,6 @@ type Pki struct {
 	path string
 }
 
-type certList struct {
-	pki           *Pki
-	serialNumbers []SerialNumber
-}
-
-type CertLister interface {
-	Count() int
-	Next() (*x509.Certificate, error)
-	All(threads int) ([]*x509.Certificate, error)
-	SerialNumbers() []SerialNumber
-}
-
 // Pki is used to return the client for PKI-related API calls.
 func (c *Client) Pki(mountPoint string) *Pki {
 	return &Pki{c: c, path: mountPoint}
@@ -78,36 +66,4 @@ func (pki *Pki) Cert(serial SerialNumber) (*x509.Certificate, error) {
 		return &x509.Certificate{}, err
 	}
 	return cert, nil
-}
-
-func (pki *Pki) List() (CertLister, error) {
-	path := fmt.Sprintf("/v1/%s/certs", pki.path)
-	r := pki.c.NewRequest("LIST", path)
-	resp, err := pki.c.RawRequest(r)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	secret, err := ParseSecret(resp.Body)
-	if err != nil {
-		return nil, err
-	}
-
-	keyInterfaces := secret.Data["keys"].([]interface{})
-	serials := make([]SerialNumber, len(keyInterfaces))
-	for i, v := range keyInterfaces {
-		s := v.(string)
-		serials[i] = SerialNumber(s)
-	}
-
-	return &certList{pki, serials}, nil
-}
-
-func (cl *certList) Count() int {
-	return len(cl.serialNumbers)
-}
-
-func (cl *certList) SerialNumbers() []SerialNumber {
-	return cl.serialNumbers
 }
